refactor(scripts): reproduce the workflow's route-type setup in `verify`
Running `verify` in a fresh worktree failed lint on an untouched route
module: `apps/web/.react-router/types` is gitignored, the workspace lint
config is type-aware, and locally those types exist only as a side effect
of `dev` or `build`. The workflow already carries a `Generate route
types` step for this; `verify` did not, so the one command meant to
reproduce CI could not.

`typegen` is now a root script that both entrypoints call, and it leads
the `verify` chain. Because it is workflow setup rather than a matrix
entry, the parity check compares the scripts named anywhere in the
workflow against the scripts `verify` chains — which is the property that
was wanted — so it is `check-verify-parity.ts` rather than
`check-verify-matrix.ts`.
