+++ package.json
@@ -11,7 +11,8 @@
     "deploy": "pnpm install --frozen-lockfile && jiti scripts/require-wrangler-config.ts && pnpm run build:web && wrangler deploy --message \"$(git rev-parse HEAD)\"",
     "db:migrate": "jiti scripts/require-wrangler-config.ts && wrangler d1 migrations apply DB",
     "db:migrate:remote": "jiti scripts/require-wrangler-config.ts && wrangler d1 migrations apply DB --remote",
-    "verify": "pnpm run lint && pnpm run typecheck && pnpm run test && pnpm run test:installers && pnpm run check:agents-md && pnpm run check:generated-assets && pnpm run check:verify-matrix && pnpm run build:web",
+    "typegen": "pnpm --filter @floway-dev/web exec react-router typegen",
+    "verify": "pnpm run typegen && pnpm run lint && pnpm run typecheck && pnpm run test && pnpm run test:installers && pnpm run check:agents-md && pnpm run check:generated-assets && pnpm run check:verify-parity && pnpm run build:web",
     "lint": "cross-env NODE_OPTIONS=\"--max-old-space-size=12288\" eslint --cache .",
     "lint:fix": "cross-env NODE_OPTIONS=\"--max-old-space-size=12288\" eslint --cache . --fix",
     "typecheck": "pnpm -r run typecheck && tsc --noEmit -p tsconfig.scripts.json",
@@ -19,7 +20,7 @@
     "test:installers": "pnpm --filter @floway-dev/agent-setup run test:installers",
     "check:agents-md": "jiti scripts/check-agents-md.ts",
     "check:generated-assets": "pnpm --filter @floway-dev/agent-setup run generate-assets --check",
-    "check:verify-matrix": "jiti scripts/check-verify-matrix.ts",
+    "check:verify-parity": "jiti scripts/check-verify-parity.ts",
     "prepare": "husky"
   },
   "devDependencies": {
